+++ src/preflightCheck.js
@@ -1,8 +1,16 @@
 import { join } from 'path';
-import assign from 'object-assign';
-import { transform } from 'babel-core';
-import { INLINE, RUNTIME, BUNDLED } from './constants.js';
-import classes from 'babel-plugin-transform-es2015-classes';
+import { transform } from '@babel/core';
+import { INLINE, RUNTIME, EXTERNAL } from './constants.js';
+
+function fallbackClassTransform () {
+	return {
+		visitor: {
+			ClassDeclaration (path, state) {
+				path.replaceWith(path, state.file.addHelper('inherits'));
+			}
+		}
+	};
+}
 
 let preflightCheckResults = {};
 
@@ -10,25 +18,28 @@ export default function preflightCheck ( options, dir ) {
 	if ( !preflightCheckResults[ dir ] ) {
 		let helpers;
 
-		options = assign( {}, options );
+		options = Object.assign( {}, options );
 		delete options.only;
 		delete options.ignore;
 
 		options.filename = join( dir, 'x.js' );
 
-		options.plugins = options.plugins ? options.plugins.concat( classes ) : [ classes ];
+		const inputCode = 'class Foo extends Bar {};\nexport default Foo;';
+		let check = transform( inputCode, options ).code;
 
-		const check = transform( 'export default class Foo {}', options ).code;
+		if ( ~check.indexOf('class ') ) {
+			options.plugins = (options.plugins || []).concat( fallbackClassTransform );
+			check = transform( inputCode, options ).code;
+		}
 
-		if ( !~check.indexOf( 'export default' ) && !~check.indexOf( 'export { Foo as default }' ) ) throw new Error( 'It looks like your Babel configuration specifies a module transformer. Please disable it. See https://github.com/rollup/rollup-plugin-babel#configuring-babel for more information' );
-
-		if ( ~check.indexOf( 'import _classCallCheck from' ) ) helpers = RUNTIME;
-		else if ( ~check.indexOf( 'function _classCallCheck' ) ) helpers = INLINE;
-		else if ( ~check.indexOf( 'babelHelpers' ) ) helpers = BUNDLED;
-
+		if ( ~check.indexOf( 'import _inherits' ) ) helpers = RUNTIME;
+		else if ( ~check.indexOf( 'function _inherits' ) ) helpers = INLINE;
+		else if ( ~check.indexOf( 'babelHelpers' ) ) helpers = EXTERNAL;
 		else {
 			throw new Error( 'An unexpected situation arose. Please raise an issue at https://github.com/rollup/rollup-plugin-babel/issues. Thanks!' );
 		}
+
+		if ( !~check.indexOf( 'export default' ) && !~check.indexOf( 'export default Foo' ) ) throw new Error( 'It looks like your Babel configuration specifies a module transformer. Please disable it. See https://github.com/rollup/rollup-plugin-babel#configuring-babel for more information' );
 
 		preflightCheckResults[ dir ] = helpers;
 	}