+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # rollup-plugin-babel changelog
 
+## 3.0.3
+
+* Drop babel7 support. Use 4.0.0-beta if you use babel 7
+* Use "module" in addition to "jsnext:main" ([#150](https://github.com/rollup/rollup-plugin-babel/issues/150))
+* Remove unused babel helpers namespace declaration & expression ([#164](https://github.com/rollup/rollup-plugin-babel/issues/164))
+
+## 3.0.2
+
+* Fix regression with Babel 6 ([#158](https://github.com/rollup/rollup-plugin-babel/issues/158))
+
+## 3.0.1
+
+* Wasn't working, fix bug with transform (not using es2015-classes for preflight check)
+
+## 3.0.0
+
+* Drop Node 0.10/0.12 (Use native `Object.assign`)
+* Change `babel-core` to be a peerDependency
+* Support `babel-core` v7 as well as a peerDep (no changes necessary)
+
 ## 2.7.1
 
 * Prevent erroneous warnings about duplicated runtime helpers ([#105](https://github.com/rollup/rollup-plugin-babel/issues/105))